+++ src/string-buffer.c
@@ -21,7 +21,7 @@ string_buffer* string_buffer_create(const size_t words_count, const size_t word_
 
     pthread_mutex_init(&buffer->mutex, NULL);
 
-    sem_init(&buffer->empty, 0, words_count);
+    sem_init(&buffer->empty, 0, (unsigned int) words_count);
     sem_init(&buffer->full, 0, 0);
 
     buffer->head = 0;
